requirements.txt
streamlit>=1.32.0
pandas>=2.2.0
numpy>=2.2.0
plotly>=5.19.0
scikit-learn>=1.4.0
requests>=2.31.0
python-dotenv>=1.0.0
fastapi>=0.110.0
uvicorn>=0.27.0
timezonefinder>=6.2.0
python-multipart>=0.0.9
geopy>=2.4.1
pydantic>=2.6.3
folium
streamlit-folium
pykalman>=0.9.5